requirements.txt
tensorflow_gpu
pyvips
pillow
pandas
matplotlib
keras_applications
scipy